Update simulation.py for issue postprocessing FEWS
FEWS expects id's in the header of the output csv-file which contains the discharges. After the 3Di update the value for the dummy-location in the output changed from 0 to an empty-string. This results in a validation error in FEWS. By setting a 0 for the dummy location this issue is resolved.